Clean up test files
